### 2021-04-23 0781625
moved LICENSE from inst/ to base directory
