### 2021-04-23 #3209 @tobiasschoch
I moved the LICENSE file from inst/ to the base directory. I added a tab in front of 'index: 1' in paper.md.
