### 2021-05-03 #3209 @fboehm
:wave: @msalibian & @aalfons & @matthias-da, would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
