### 2021-05-03 #3209 @msalibian
@fboehm is there a way to change my settings so I only get emails from this issue? I seem to be currently receiving email notifications from many other joss-reviews/issues that are not relevant to me.
