### 2021-05-04 #3209 @fboehm
@msalibian - I'm not sure why that is. Here is a link that explains some ways that this problem could arise: https://github.com/openjournals/joss/issues/200

Maybe start by trying the suggested solutions there. If they don't work, please let me know. I can reach out to other JOSS editors if the problem persists.
