### 2021-05-04 #3238 @whedon
[REVIEW]: wbacon: Weighted BACON algorithms for multivariate outlier nomination (detection) and robust linear regression
**Submitting author:** @tobiasschoch (<a href="http://orcid.org/0000-0002-1640-3395">Tobias Schoch</a>)
**Repository:** <a href="https://github.com/tobiasschoch/wbacon" target ="_blank">https://github.com/tobiasschoch/wbacon</a>
**Version:** v0.5
