### 2021-05-13 #3238 @msalibian
This is a nice contribution (both package and manuscript). I have some minor suggestions about the text, and noted that some items from the checklist above are missing.

- In the text, when referring to "large datasets", please be specific if this is regarding the number of cases (rows), variables (columns), or both. In the robustness literature, computational effort ("complexity" in an informal way) usually grows exponentially with the number of columns, while a larger number of cases does not generally increase the cost faster than linearly.
- There are references to the "BACON algorithms", it may be helpful for the reader if the Summary mentioned that the two algorithms correspond to either a multivariate location/scatter model, or a linear regression one.
- line 12: "superior" is a relative term that requires a comparison (superior to what?), I suggest simply mentioning the method's popularity instead.
- line 15: I believe the preferred term in `R` is "package" (rather than "library").
- line 18: "etc.", I suggest listing all the implemented diagnostic methods / tools instead.
- "Example usage": currently missing. It would be very useful to include an example illustrating the usage of the package (including how to set tuning parameters), and use it also to compare its performance against that of other BACON implementations.
- "State of the field" and  "Performance": although the author describes how this software compares to other implementations, a detailed example would be very useful (as I mention above). In particular, it is not currently clear if the "large datasets" for which this package is geared are "large" in terms of the number of cases (rows), the number of variables (columns), or both.
- "Community guidelines" are missing.
