### 2021-05-14 #3238 @tobiasschoch
@msalibian - Thank you for the very helpful remarks. Before I’m going to address the points raised, I have some questions for the editor.

@fboehm - I have the following questions:

1. I’m planning to do the following (while working through the issues raised by the reviewers):
 - Make changes in the documents/ code (in the repository)
 - Add a new version tag to the repository
 - Call [at]whedon set [new version] as version
 - Call [at]whedon generate pdf
 - **Is that right?**

2. “Example usage”: The package has a vignette. In the vignette, the methods are applied to example data. **Should I include the examples or some of the examples in the paper as well? Or should I just note in the paper that there is a vignette?**

3. “Community guidelines”: These guidelines can be found in the README.md on https://github.com/tobiasschoch/wbacon. I put them there such that they show up on the GitHub site (this is the place where most people will “land” while browsing/ searching the internet). **Should I include the “community guidelines” in the paper as well?**

4. "Performance":
 - I test my implementation against (the existing code) <code>robustX::BACON</code>; see folder test (focus: does my implementation give the same results?).
 - In addition, there are some benchmarks; see test/benchmark (used to “see” the impact of the OpenMP parallelization).
 - However, there are yet no benchmarks, where I compare my implementation against <code>robustX::BACON</code> in terms of performance/ speed. I’m planning to add such a benchmark. Then, I note in the paper that such benchmark is available. **Is that ok?**
