### 2021-05-14 #3238 @fboehm
@tobiasschoch - Thanks for your questions. Below I try to answer them.

1. Don't worry about making the tagged release or interacting with whedon. I'll guide you through that when it's time. For now, just concentrate on your first bullet point. If you want to see how the pdf looks after your changes are made, you can tell whedon to "generate pdf", but don't create a new release until I ask you to do so.

2. I'm not sure about this one. I think the paper.md might be better off with a mention of the vignette. You probably don't need to include examples in the paper.md file. @msalibian  - do you have thoughts on this?

3. I think that including guidelines in the README.md is sufficient. However, you might also want to add a CONTRIBUTING.md file to the repository. For example, see this: https://github.com/tidyverse/dplyr/blob/master/.github/CONTRIBUTING.md. You don't need to include community guidelines in the paper.md. There is a function in the R package `usethis` that can create such a file from a template: `usethis::use_tidy_contributing`. You might want to edit the resulting file.

4. Yes, I think that mentioning in paper.md that you have the tests and benchmarks in the package is sufficient. You might also include unit tests, possibly by using the R package `testthat`. @msalibian - do you agree? Do you have more specific comments to add on this point?
