### 2021-05-17 #3238 @tobiasschoch
@fboehm @msalibian @aalfons

 - I'm greatful for all the hints and comments. Now I understand how the review process works at JOSS.
 - Frederick, I don't mind if the review takes more time than usual. The reviewers are busy professors (so am I). I prefer a detailed review to a quick review. We should give aalfons more time if he needs it.
 - Next, I'm going to address the points raised by msalibian.
