+++ inst/paper/paper.md
@@ -14,31 +14,29 @@ authors:
 affiliations:
  - name: University of Applied Sciences and Arts Northwestern Switzerland, School of Business, Riggenbachstrasse 16, CH-4600 Olten, Switzerland
    index: 1
-date: 29 March 2021
+date: "16 May 2021"
 bibliography: paper.bib
 ---
 
 # Summary
+Outlier nomination (detection) and robust regression are computationally hard problems. This is all the more true when the number of variables and observations grow rapidly. Among all candidate methods, the two BACON (blocked adaptive computationally efficient outlier nominators) algorithms of @billoretal2000 have favorable computational characteristics as they require only a few model evaluations irrespective of the sample size. This makes them popular algorithms for multivariate outlier nomination/ detection and robust linear regression (at the time of writing [Google Scholar](https://scholar.google.com) reports more than 500 citations of the @billoretal2000 paper).
 
-Outlier nomination (detection) and robust regression are computationally hard problems. This is all the more true when the number of variables and observations grow rapidly. Among all candidate methods, the BACON (blocked adaptive computationally efficient outlier nominators) algorithms of @billoretal2000 have favorable computational characteristics as they require only a few model evaluations irrespective of the sample size. This makes them superior and popular algorithms (at the time of writing [Google Scholar](https://scholar.google.com) reports more than 500 citations of the @billoretal2000 paper).
-
-`wbacon` is a package/library for the `R` statistical software [@r2021]. The library is aimed at medium to large datasets that can possibly have (sampling) weights (e.g., data from complex survey samples). The library has a user-friendly `R` interface (with plotting methods, etc.) and is written mainly in the `C` language (with OpenMP support for parallelization; see @openmp2018) for performance reasons.
+`wbacon` is a package for the `R` statistical software [@r2021]. It is aimed at medium to large datasets that can possibly have (sampling) weights (e.g., data from complex survey samples). The package has a user-friendly `R` interface (with plotting methods, etc.) and is written mainly in the `C` language (with OpenMP support for parallelization; see @openmp2018) for performance reasons.
 
 # The BACON algorithms
-
 Technically, the BACON algorithms consist of the application of series of simple statistical estimation methods such as coordinate-wise means/medians, covariance matrix, Mahalanobis distances, or least squares regression on subsets of the data. The algorithms start from an initial small subset of "good" data and keep adding those observations to the subset whose distances/ discrepancies are smaller than a predefined threshold value. The algorithms terminate if the subset cannot be increased further. The observations not in the final subset are nominated as outliers. We follow @billoretal2000 and use the term "nomination" of outliers instead of "detection" to emphasize that the algorithms should not go beyond nominating observations as potential outliers. It is left to the analyst to finally label outlying observations as such.
+
[…]
