+++ inst/paper/paper.bib
@@ -19,6 +19,16 @@
     doi = {10.1016/S0167-9473(99)00101-2}
 }
 
+@book{chatterjeehadi2006,
+    title = {Regression Analysis by Example},
+    publisher = {John Wiley and Sons},
+    year = {2006},
+    author = {S. Chatterjee and A. H. Hadi},
+    address = {Hoboken (NJ)},
+    edition = {4th},
+    doi = {10.1002/0470055464}
+}
+
 @article{blas2002,
     author = {L. S. Blackford and A. Petitet and R. Pozo and K. Remington
         and R. C. Whaley and J. Demmel and J. Dongarra and I. Duff and
@@ -51,6 +61,15 @@
     address = {Philadelphia},
     edition = {3rd},
     doi = {10.1137/1.9780898719604}
+}
+
+@manual{microbenchmark,
+    author = {O. Mersmann, C. Beleites, R. Hurling, A. Friedman and
+        J. M. Ulrich},
+    title = {microbenchmark: Accurate Timing Functions},
+    year = {2019},
+    organization = {{R} package version 1.4-7},
+    url = {https://CRAN.R-project.org/package=microbenchmark}
 }
 
 @manual{modi,
@@ -87,6 +106,16 @@
     address = {Vienna, Austria},
     year = {2021},
     url = {https://www.R-project.org/},
+}
+
+@manual{robustbase,
+    author = {M. Maechler, P. Rousseeuw, C. Croux, V. Todorov,
+        A. Ruckstuhl, M. Salibian-Barrera, T. Verbeke, M. Koller,
+        E. L. T. Conceicao, M. A. di Palma},
+    title = {robustbase: Basic Robust Statistics},
+    year = {2021},
+    organization = {{R} package version 0.93-7},
+    url = {https://CRAN.R-project.org/package=robustbase}
 }
 
 @manual{robustx,
