+++ inst/paper/paper.bib
@@ -64,7 +64,7 @@
 }
 
 @manual{microbenchmark,
-    author = {O. Mersmann, C. Beleites, R. Hurling, A. Friedman and
+    author = {O. Mersmann and C. Beleites and R. Hurling and A. Friedman and
         J. M. Ulrich},
     title = {microbenchmark: Accurate Timing Functions},
     year = {2019},
@@ -109,9 +109,9 @@
 }
 
 @manual{robustbase,
-    author = {M. Maechler, P. Rousseeuw, C. Croux, V. Todorov,
-        A. Ruckstuhl, M. Salibian-Barrera, T. Verbeke, M. Koller,
-        E. L. T. Conceicao, M. A. di Palma},
+    author = {M. Maechler and P. Rousseeuw and C. Croux and V. Todorov and
+        A. Ruckstuhl and M. Salibian-Barrera and T. Verbeke and M. Koller and
+        E. L. T. Conceicao and M. A. di Palma},
     title = {robustbase: Basic Robust Statistics},
     year = {2021},
     organization = {{R} package version 0.93-7},
