+++ inst/paper/paper.md
@@ -26,15 +26,15 @@ Outlier nomination (detection) and robust regression are computationally hard pr
 # The BACON algorithms
 Technically, the BACON algorithms consist of the application of series of simple statistical estimation methods such as coordinate-wise means/medians, covariance matrix, Mahalanobis distances, or least squares regression on subsets of the data. The algorithms start from an initial small subset of "good" data and keep adding those observations to the subset whose distances/ discrepancies are smaller than a predefined threshold value. The algorithms terminate if the subset cannot be increased further. The observations not in the final subset are nominated as outliers. We follow @billoretal2000 and use the term "nomination" of outliers instead of "detection" to emphasize that the algorithms should not go beyond nominating observations as potential outliers. It is left to the analyst to finally label outlying observations as such.
 
-The BACON algorithm for multivariate outlier nomination can be initialized in two ways: version "V1" or "V2" [see @billoretal2000]. In version "V1", the algorithm is started from the coordinatewise median. As a consequence, the resulting estimators of location and scatter are robust (the breakdown point is approximately 40\%, see @billoretal2000) but not affine equivariant estimators of the population location and scatter. However, @billoretal2000 show that the estimators are *nearly* affine equivariant. The initialization by version "V1" leads to estimators that are affine equivariant by design because the algorithm is started from the coordinatewise mean but the estimators have a very low breakdown point.
+The BACON algorithm for multivariate outlier nomination can be initialized in two ways: version "V1" or "V2" [see @billoretal2000]. In version "V1", the algorithm is started from the coordinatewise median. As a consequence, the resulting estimators of location and scatter are robust (the breakdown point is approximately 40\%, see @billoretal2000) but not affine equivariant estimators of the population location and scatter. However, @billoretal2000 show that the estimators are *nearly* affine equivariant. The initialization by version "V1" yields estimators that are affine equivariant by design because the algorithm is started from the coordinatewise mean but the estimators have a very low breakdown point.
 
 A naive implementation of the BACON algorithms would call the (simple) estimation methods iteratively on a sequence of growing subsets of the data without bothering too much with reusing or updating existing blocks of data. This leads to an excessively large number of copy/ modify operations and (unnecessary) recomputations. Overall, we would end up with a computationally inefficient implementation. With small datasets, the inefficiencies would probably not be noticed. With large amounts of data, however, the situation is quite different.
 
 # Statement of need
-The two BACON algorithms are available from the package `robustX` [@robustx] for the `R` statistical software. The BACON algorithm for multivariate outlier nomination has been extended  to weighted problems (in the context of survey sampling) and incomplete data by @beguinhulliger2008. The extended method is available from the `R` package `modi` [@modi]. Both implementations are not explicitly targeted at big data applications. Our package fills this gap.
[…]
