### 2021-05-19 60c753b
cellWise::philips changed name to data_philips
The name change in cellWise affects the vignette
