+++ inst/paper/paper.md
@@ -132,7 +132,7 @@ See vignette to learn more about the package.
 
 We benchmark our implementation against `robustX::BACON()`. Our implementation is intentionally *limited to single-threaded* computations (`n_threads = 1`; no OpenMP parallelization support). It is clear that when the number of variables is large, the parallelized computations are (usually) much faster.
 
-We consider estimating a robust linear regression for a Gaussian mixture distribution, where a proportion of $1- \epsilon$ of the observations on the $p$ independent variables is generated by the Gaussian model, while a proportion of $\epsilon$ (the outliers) is generated by a shifted Gaussian distribution. For the outlying observations (i.e., $\epsilon$ proportion of the data), the response variable is generated by a regression coefficient which is 10 times larger than the coefficient of the non-outlying observations. We chose $\epsilon = 0.05$; see Appendix for more details on the setup. The number of variables ($p$) and the number of observations ($n$) are varied. Table 1 shows the ratio of average compute time of the two implementation for some configurations of $(n,p)$. A ratio $> 1.0$ ($< 1.0$) implies that `wBACON_reg()` is faster (slower) than `robustX:BACON()`. The average ratio refers to compute time averages over repeated benchmarks using the R package `microbenchmark` [@microbenchmark]. It is evident from the results in Table 1 that `wBACON_reg()` is considerably faster than its competitor, e.g.,`wBACON_reg()` is on average 4.4 times faster for the setup $p=5$ and $n=100$. More importantly, the differences become larger as we increase $n$ or $p$. The differences in compute speed are mainly due to the fact that `wBACON_reg()` updates the regression estimates as the subset of non-outlying observations grows, while `robustX:BACON()` recomputes the estimates at each iteration.
+We consider estimating a robust linear regression for a Gaussian mixture distribution, where a proportion of $1- \epsilon$ of the observations on the $p$ independent variables is generated by the Gaussian model, while a proportion of $\epsilon$ (the outliers) is generated by a shifted Gaussian distribution. For the outlying observations (i.e., $\epsilon$ proportion of the data), the response variable is generated by a regression coefficient which is 10 times larger than the coefficient of the non-outlying observations. We chose $\epsilon = 0.05$; see Appendix for more details on the setup. The number of variables ($p$) and the number of observations ($n$) are varied. Table 1 shows the ratio of average compute time of the two implementation for some configurations of $(n,p)$. A ratio $> 1.0$ ($< 1.0$) implies that `wBACON_reg()` is faster (slower) than `robustX:BACON()`. The average ratio refers to compute time averages over repeated benchmarks using the R package `microbenchmark` [@microbenchmark]. It is evident from the results in Table 1 that `wBACON_reg()` is considerably faster than its competitor, e.g., `wBACON_reg()` is on average 4.4 times faster for the setup $p=5$ and $n=100$. More importantly, the differences become larger as we increase $n$ or $p$. The differences in compute speed are mainly due to the fact that `wBACON_reg()` updates the regression estimates as the subset of non-outlying observations grows, while `robustX:BACON()` recomputes the estimates at each iteration.
 
 | No. of variables $p$ | No. of observations $n$ | Ratio |
[…]
