### 2021-05-20 3e5fe6e
JOSS paper: benchmark
