+++ inst/paper/paper.md
@@ -143,53 +143,21 @@ We consider estimating a robust linear regression for a Gaussian mixture distrib
 | 10 | 1000 | 52.5 |
 | 20 | 1000 | 55.0 |
 
-Table 1: Benchmarks: Robust linear regression
-
-
-| No. of variables | No. of observations | Ratio |
-| :--------------- | :------------------ | ----: |
-|  5 |    1,000 | 5.9 |
-| 10 |    1,000 | 3.9 |
-| 20 |    1,000 | 2.5 |
-|  5 |   10,000 | 6.5 |
-| 10 |   10,000 | 3.7 |
-| 20 |   10,000 | 2.8 |
-|  5 |  100,000 | 7.3 |
-| 10 |  100,000 | 4.2 |
-| 20 |  100,000 | 3.3 |
-|  5 | 1,000,000 | 6.2 |
-| 10 | 1,000,000 | 4.1 |
-| 20 | 1,000,000 | 3.2 |
-
-Table 2: Benchmarks: Multivariate outlier nomination/ detection
+Table 1: Benchmarks: Robust linear regression (single-threaded)
 
 
 
-| No. of variables | No. of observations | Ratio |
-| :--------------- | :------------------ | ----: |
-|  5 |    1,000 | 5.9 |
-| 10 |    1,000 | 3.9 |
-| 20 |    1,000 | 2.5 |
 
-|  5 |   10,000 | 6.5 |
-| 10 |   10,000 | 3.7 |
-| 20 |   10,000 | 2.8 |
+|            | $p=5$ | $p=10$ | $p=20$ | $p=30$ | $p=40$ | $p=50$ | $p=100$ | $p=200$ |
+| :--------- | ----: | -----: | -----: | -----: | -----: | -----: | ------: | ------: |
+| $n = 10^3$ | 0.5 | 1.0 | 1.2 | 1.4 | 2.3 | 2.7 | 4.7 |  9.0 |
+| $n = 10^4$ | 0.9 | 1.1 | 1.5 | 1.7 | 2.6 | 3.0 | 5.0 |  8.3 |
+| $n = 10^5$ | 0.1 | 0.3 | 0.8 | 1.3 | 2.0 | 2.3 | 4.7 |  9.9 |
+| $n = 10^6$ | 0.9 | 1.5 | 2.2 | 2.8 | 4.0 | 4.1 | 7.7 | 13.7 |
 
-|  5 |  100,000 | 7.3 |
-| 10 |  100,000 | 4.2 |
-| 20 |  100,000 | 3.3 |
-
-|  5 | 1,000,000 | 6.2 |
-| 10 | 1,000,000 | 4.1 |
-| 20 | 1,000,000 | 3.2 |
+Table 2: Benchmarks: Multivariate outlier nomination/ detection (multi-threaded)
 
 
-| n         |  p = 5 | p = 10 | p = 20 |
-| :-------- | -----: | -----: | -----: |
-|     1,000 | 5.9 | 3.9 | 2.5 |
-|    10,000 | 6.5 | 3.7 | 2.8 |
-|   100,000 | 7.3 | 4.2 | 3.3 |
-| 1,000,000 | 6.2 | 5.1 | 3.2 |
 
 
 # Community guidelines
