### 2021-05-21 7bdf224
JOSS paper: update 3; modified benchmark
