+++ tests/benchmark/benchmark_robustx.R
@@ -13,18 +13,15 @@ if (robustX_version < 1.25) {
         robustX_version, "\nGrab the new version from R-Forge\n"))
 }
 library(microbenchmark)
-
-#*******************************************************************************
-# Important Note: The functions wbacon::wBACON and wbacon::wBACON_reg have
-#                 OpenMP (parallelization) support. The parallelization is over
-#                 the number of variables p. Thus, these functions can easily
-#                 outperform the functions in package robustX when p is large.
-#                 In order to have a 'fair' comparison, the OpenMP support is
-#                 turned off (see argument 'n_threads = 1').
-#*******************************************************************************
+#===============================================================================
+# 0 SETUP
+#===============================================================================
+VERSION   <- "V2"    # Initialization (version "V1" or "V2")
+ALPHA     <- 0.05    # Cutoff value: (1-ALPHA) quantile of the chi square distr.
+N_THREADS <- 1       # OpenMP number of threads
 
 #===============================================================================
-# I. BACON robust linear regression
+# 1 BACON robust linear regression
 #===============================================================================
 # scenarios
 #-------------------------------------------------------------------------------
@@ -46,9 +43,6 @@ colnames(scenario) <- c("p", "n", "eps", "times")
 #-------------------------------------------------------------------------------
 # microbenchmark
 #-------------------------------------------------------------------------------
-VERSION <- "V2" # Initialization (version "V1" or "V2")
-ALPHA <- 0.05   # Cutoff value: (1-ALPHA) quantile of the chi square distr.
-
 scenario$ratio <- NA
 for (i in 1:NROW(scenario)) {
     p <- scenario[i, "p"]
@@ -71,7 +65,7 @@ for (i in 1:NROW(scenario)) {
     m0 <- microbenchmark(BACON(x, y, alpha = ALPHA, init.sel = VERSION,
         verbose = FALSE), times = scenario$times[i])
     m1 <- microbenchmark(wBACON_reg(y ~ ., data = dat, alpha = ALPHA,
-        version = VERSION, n_threads = 1), times = scenario$times[i])
+        version = VERSION, n_threads = N_THREADS), times = scenario$times[i])
     scenario$ratio[i] <- mean(m0$time) / mean(m1$time)
 
     cat(i, "of", NROW(scenario), "\n")
@@ -83,26 +77,46 @@ for (i in 1:NROW(scenario)) {
 scenario
 
 #===============================================================================
-# II. BACON multivariate outlier nomination (detection)
+# 2 BACON multivariate outlier nomination (detection)
 #===============================================================================
 # scenarios
 #-------------------------------------------------------------------------------
 scenario <- data.frame(matrix(c(
-#---------------------------------
-#       p         n   eps   times
-#---------------------------------
-       5,      1000,  0.05,    50,
-       10,     1000,  0.05,    50,
-       20,     1000,  0.05,    50,
-       5,     10000,  0.05,    30,
-       10,    10000,  0.05,    30,
-       20,    10000,  0.05,    30,
-       5,    100000,  0.05,    10,
-       10,   100000,  0.05,    10,
-       20,   100000,  0.05,    10,
-       5,   1000000,  0.05,    2,
-       10,  1000000,  0.05,    2,
-       20,  1000000,  0.05,    2),
+#------------------------------------
+#         p        n     eps  times
+#------------------------------------
[…]
