+++ inst/doc/doc_c_functions.tex
@@ -116,7 +116,7 @@
 	\end{center}
 \end{minipage}}
 
-\date{{\small \today, version 0.3}}
+\date{{\small \today}}
 \maketitle
 
 \setstretch{0.8}

[…]
