### 2021-05-21 #3238 @tobiasschoch
I modified the paper according to @msalibian  suggestions (I'm going to address the points raised by @aalfons later)

### Issue 1

 - *Reviewer:* In the text, when referring to "large datasets", please be specific if this is regarding the number of cases (rows), variables (columns), or both. In the robustness literature, computational effort ("complexity" in an informal way) usually grows exponentially with the number of columns, while a larger number of cases does not generally increase the cost faster than linearly.
 - *Answer:* In the revised manuscript, I now point out that the time complexity of the algorithms is dominated by the number of variables/ columns. Therefore, the implemented parallelization over the columns is meaningful. See also the new chapter on benchmarking.

### Issue 2

 - *Reviewer:* There are references to the "BACON algorithms", it may be helpful for the reader if the Summary mentioned that the two algorithms correspond to either a multivariate location/scatter model, or a linear regression one.
 - *Answer:* Yes, I now mention the two methods.

### Issue 3

 - *Reviewer:* line 12: "superior" is a relative term that requires a comparison (superior to what?), I suggest simply mentioning the method's popularity instead.
 - *Answer:* Yes, I deleted the ambiguous term “superior”.

### Issue 4

 - *Reviewer:* line 15: I believe the preferred term in R is "package" (rather than "library").
 - *Answer:*  Yes, I now talk about “package” not “library”.

### Issue 5

 - *Reviewer:*  line 18: "etc.", I suggest listing all the implemented diagnostic methods / tools instead.
 - *Answer:* I chose to stick with the (rather unspecific) “etc.” instead of listing all methods as this would have made the summary bulky. However, I now added a chapter called “Illustration” where I show an application of some of the methods and refer the reader to the vignette to learn more.

### Issue 6

- *Reviewer:* "Example usage": currently missing. It would be very useful to include an example illustrating the usage of the package (including how to set tuning parameters), and use it also to compare its performance against that of other BACON implementations.
 - *Answer:* I added a small example on robust regression. This is meant as a kind of teaser to get readers interested in the vignette.

### Issue 7

- *Reviewer:* "State of the field" and "Performance": although the author describes how this software compares to other implementations, a detailed example would be very useful (as I mention above). In particular, it is not currently clear if the "large datasets" for which this package is geared are "large" in terms of the number of cases (rows), the number of variables (columns), or both.
- *Answer:* I added a chapter on “benchmarking”, where I compare my implementation with the one in the robustX package (reference) in terms of computation time for data sets in various sizes. From the benchmarking, we can learn that my implementation outperforms the reference on medium to large data sets.

### Issue 8

 - *Reviewer:* "Community guidelines" are missing.
 - *Answer:* I already had the file “CONTRIBUTING.md” in the base directory of the package. Now I added a chapter on "community guidelines" to the paper.
