### 2021-05-21 e6dbf83
bugfix: include guard: OpenMP support for Apple MacOS
