### 2021-05-21 #3238 @tobiasschoch
@aalfons - Thank you for the very helpful remarks. Your comments reached me while I was responding to the points raised by msalibian. I will now address the points you have raised.

### Issue 1
*Reviewer:* **Installation instructions:** I received the following error message when trying to install the package [...].

*Answer:* Yes, indeed I forgot to add the (conditional) include guards. This is now fixed. I added the include guards in the C header files.

### Issue 2

*Reviewer:* **Performance**: line 36: What size of the data set (number of observations, number of variables) is approximately necessary for implementation inefficiencies to matter?

*Answer:* In the mean time I have updated the paper. In the "Benchmarking" section, I compare my implementation with the reference implementation `robustX` in terms of computation time for data sets in various sizes. From the benchmarking, we can learn that my implementation outperforms the reference on medium to large data sets.

### Issue 3

*Reviewer:* **Automated tests**: There are some scripts in the tests/ directory that seem to reproduce the results of a paper and contain some speed tests of the computational performance, but I did not find any unit tests that the functions produce the correct output and so on. On the other hand, the documentation contains some examples, and it can be automatically checked whether these produce warnings or errors via R CMD check. So in that sense, there would be some minimal automated testing of the functionality. @fboehm Could you please provide some guidance on this point?

*Answer:* I am a believer in test-driven development. I distinguish between two (idealized) types of tests:
 - **Unit test:** *Test what the code is*. The purpose of unit testing is to isolate the smallest testable parts of the code base and verify whether they function properly in isolation.
 - **Functional test:** *Test what the code does*. A tester is not concerned with the actual code, rather s/he wants to verify the output with the expected output.

As a pragmatic programmer I take the following view: Some functions are suitable for unit testing, others less so.
 - I do have unit tests for the functions `wquantile` and `wselect`. Because these functions are also used in other R packages, they reside in their own GitHub repo called "wquantile". There, you can find unit tests based on the C language CHEAT test library of Guillermo Freschi and Sampsa Kiiskinen; see https://github.com/tobiasschoch/wquantile/blob/master/tests/test.c.
 - The vast majority of code in the "wbacon" repo is tested under the "functional testing" paradigm for the following reason: The BACON algorithms are based on the basic "building blocks": mean, covariance matrix, linear regression, etc. In my package, these building blocks are computed with the help of the LAPACK and BLAS subroutines. More precisely,

    - the covariance matrix is based on `dsyrk` => computes the lower triangular scatter matrix (see wbacon.c).
    - the regression coefficients are computed with `dgels`; the residuals are computed with `dgemv` (see fitwls.c).
    - ...

We could, in principle, follow the unit-test paradigm and write separate unit tests for the covariance matrix, regression, and so on. However, this would *ultimately* boil down to testing the subroutines in BLAS and LAPACK. Neither is my package the place to test these subroutines, nor would such a testing strategy increase confidence in my functions. Of course, that does not rule out the possibility that I mixed up the order of the arguments when calling the subroutines. But the compiler (and valgrind) would have told me if I had actually done it. So I decided to take a "functional test" approach and test the *ensemble* of building blocks. In total, the BACON algorithms are tested on 159 different (real) test sets to match the results of the (reference implementation) in the `robustX` package.

### Issue 4

*Reviewer:* **State of the field**: The authors mention only other packages that implement the same algorithms. Maybe it would be good to also mention some other popular packages that provide other algorithms for anomaly detection? (Although this could quickly go out of hand, as there are many algorithms for this purpose.)

*Answer:* Yes, this could indeed go out of hand quickly... I focused on the BACON algorithms and did not intend to write a review article on multivariate outlier detection. @fboehm: Shall I discuss other methods?

### Issue 5

*Reviewer:* line 12: the statement that the BACON algorithms are "superior" would imply that they are always better than other anomaly detection algorithms. Please rephrase this.

*Answer:* Yes, this has been fixed; see response to points raised by msalibian.

### Issue 6

*Reviewer:* lines 15-17: Even though the function to load a package is called library(), which causes some confusion about the terminology, the preferred term is "package" in R. Please change this throughout the paper.

*Answer:* Yes, this has been fixed; see response to points raised by msalibian.

### Issue 7

*Reviewer:* line 66/67: How should one proceed if the diagnostic tools indicate that the "good" observations violate the structure that is required by the algorithm?

*Answer:* Well...  all I can do (see also vignette) is to quote the developpers of the BACON algorithms: “Although the algorithms will often do something reasonable even when these assumptions are violated, it is hard to say what the results mean.” Billor et al. (2000, p. 290). It is better to be "safe than sorry" and apply another method. I think I have made it sufficiently clear when the method can and cannot be used; see vignette.
