### 2021-05-31 #3238 @msalibian
@fboehm Thanks @tobiasschoch for addressing my concerns and suggestions. I have no pending items in the check list. I am satisfied with the new version of the manuscript, and only have a few minor suggestions about writing style, which I leave here in case they are helpful.

- The quote on lines 80-82 seems out of place, there is no reference to it in the text either before or after it;
- Line 83, "tools to study potentially outlying observations", I'd say "tools to identify potentially...";
- Line 96, "we study the BACON algorithm", I'd instead say: "we illustrate the use of the BACON algorithm";
- Line 116, "distances / discrepancies", are these distances (as in Mahalanobis distances), or standardized residuals? The fact that the reference distribution is a Student's T makes me suspect they are the latter. It would be useful to make the text as precise as possible;
- Lines 138-139, it'd be appropriate to include here that this plot is also included in the `plot` method for `lmrob` objects in package `robustbase`;
- Line 162: `wBACON_reg_reg()`, should this be `wBACON_reg()` instead?
- Line 177, "easily outperforms", a less informal phrase would be to just use "outperforms", or "clearly outperforms", etc.;
- Lines 188-189, I'm not sure it is necessary to include in the manuscript the suggestion to search existing issues in the software repository before submitting a new one.
