### 2021-05-31 #3238 @aalfons
@fboehm I'm also satisfied with the new version of the manuscript and have no more unchecked items. Thanks @tobiasschoch for addressing all the comments.

I also have a few minor suggestions regarding the text to add to those of @msalibian which I leave here in the same spirit in case they are helpful:

- Line 16: "R statistical software" -> "statistical software R"
- Line 168: "... of the project GitHub repository" - it might be useful to include the link here.
- Line 208: "Compute environment" -> "Computing environment"
