+++ inst/paper/paper.md
@@ -88,7 +88,7 @@ reg
 
 By default, `wBACON_reg()` uses the parametrization `alpha = 0.05`, `collect = 4`, and `version = "V2"`. These parameters are used to call the `wBACON()` multivariate outlier nomination/ detection algorithm on the design matrix. Then, the same parameters are used to compute the robust linear regression.
 
-To ensure a high breakdown point, `version = "V2"` should not be changed to "V1" unless you have good reasons to do so. The main "turning knob" to tune the regression algorithm is `alpha`, which defines the $(1 - \alpha)$ quantile of the Student $t$-distribution. All observations whose discrepancies (defined as the scaled residuals on the set of "good" observations and the scaled prediction error on the set of "bad" observations) are smaller (in absolute value) than the quantile are selected into the subset of "good" data [see document `methods.pdf` in the folder `inst/doc` of the source package]. By choosing larger values for `alpha` (e.g., 0.2), more observations are selected (ceteris paribus) into the subset of "good" data (and vice versa).
+To ensure a high breakdown point, `version = "V2"` should not be changed to "V1" unless you have good reasons to do so. The main "turning knob" to tune the regression algorithm is `alpha`, which defines the $(1 - \alpha)$ quantile $t_{\alpha,\nu}$ of the Student $t$-distribution with $\nu$ degrees of freedom. In fact, the quantile $t_{\alpha/(2r + 2),r-p}$ is used as the cutoff value [see @billoretal2000], where $r$ and $p$ denote, respectively, the number observations in the set of "good" observations and the number of variables. All observations whose discrepancies (defined as the scaled residuals on the set of "good" observations and the scaled prediction error on the set of "bad" observations) are smaller (in absolute value) than the cutoff value are selected into the subset of "good" data [see document `methods.pdf` in the folder `inst/doc` of the source package]. By choosing larger values for `alpha` (e.g., 0.2), more observations are selected (ceteris paribus) into the subset of "good" data (and vice versa).
 
 The parameter `collect` specifies the size of the initial subset, which is defined as $m = p \cdot collect$. It should be chosen such that $m$ is considerably smaller than the number of observations $n$. Otherwise we are at risk of selecting too many "bad" observations into the initial subset, which will eventually bias the regression estimates.
 
