### 2021-05-31 #3238 @tobiasschoch
@fboehm @aalfons @msalibian - Thank you for the very helpful comments (and the typos you spotted). I have followed all your suggestions and modified the paper accordingly.

@msalibian - I have clarified the statement about the "distances / discrepancies". The BACON algorithm for robust regression uses what Billor et al. (2000) call discrepancies: that is, on the set of "good" observations (in the current iteration), the discrepancies are defined as the scaled residuals. For the "bad" observations, the discrepancies are taken to be the scaled (out-of-sample) prediction error. Then, for the next iteration, all observations whose discrepancies are smaller (in absolute value) than the cutoff value (which is defined as a rather "special" Student t-quantile) are selected into the new set of "good" observations. This procedure is repeated until convergence. From the point of rigorous statistical reasoning, the t-quantile cutoff value suggested by Billor et al. (2000) is weird or questionable... but it works well in practice...
