### 2021-06-01 #3238 @fboehm
Before we proceed, I have a few small suggestions for the paper.md.

1. Lines 78-79: This sentence feels awkward. Can you rewrite it to flow a little better while still conveying the message? Perhaps something like this: "We recommend that the user examine both the data structure and the "good" observations to verify that the assumptions hold."
2. Line 84: Replace "library" with "package".
3. Section "Community guidelines": I think that this should be removed from the manuscript. I erred in overlooking your above  discussion about this. I apologize for that. What you currently have in the README.md is sufficient.
4. Just a point to consider - it seems that you don't define "breakdown point". Is this something that you want to define in the manuscript?

Once you've addressed these minor points, we will proceed towards publication. Thanks again, @tobiasschoch !
