+++ inst/paper/paper.bib
@@ -63,6 +63,17 @@
     doi = {10.1137/1.9780898719604}
 }
 
+@book{maronnaetal2018,
+    title = {Robust Statistics: Theory and Methods (with R)},
+    publisher = {John Wiley and Sons},
+    year = {2018},
+    author = {R. A. Maronna and R. D. Martin and V. J. Yohai and
+        M. Salibi{\'a}n-Barrera},
+    address = {Hoboken (NJ)},
+    edition = {2nd},
+    doi = {10.1002/9781119214656}
+}
+
 @manual{microbenchmark,
     author = {O. Mersmann and C. Beleites and R. Hurling and A. Friedman and
         J. M. Ulrich},
