+++ inst/paper/paper.bib
@@ -64,7 +64,7 @@
 }
 
 @book{maronnaetal2018,
-    title = {Robust Statistics: Theory and Methods (with R)},
+    title = {Robust Statistics: Theory and Methods (with {R})},
     publisher = {John Wiley and Sons},
     year = {2018},
     author = {R. A. Maronna and R. D. Martin and V. J. Yohai and
