### 2021-06-02 #3238 @tobiasschoch
@fboehm - Thank you for proofreading the paper. Highly appreciated!

 1.  Lines 78-79: This sentence feels awkward. Can you rewrite it to flow a little better while still conveying the message? Perhaps something like this: "We recommend that the user examine both the data structure and the "good" observations to verify that the assumptions hold." **Answer:** Yes, I have modified the sentence.
 2.  Line 84: Replace "library" with "package". **Answer:** Yes, I have replaced "library" with "package".
 3 . Section "Community guidelines": I think that this should be removed from the manuscript. I erred in overlooking your above discussion about this. I apologize for that. What you currently have in the README.md is sufficient. **Answer:** No problem. I have removed the Section "Community guidelines" from the paper.
 4. Just a point to consider - it seems that you don't define "breakdown point". Is this something that you want to define in the manuscript? **Answer:** Thank you for pointing this out. I have added a footnote that explains what the breakdown point is.

See new pdf above this post. In addition, I updated the package version.  Do I need to call `[at]whedon set [new version] as version` now?
