### 2021-06-05 #3238 @Kevin-Mattheus-Moerman
@tobiasschoch

- [ ] Can you ensure the author information of the Zenodo archive version is complete and matches that of the paper? Please use your full name and you should be able to add your ORCID id there too.
- [ ] In the paper when using the `/` it appears the authors sometimes follow it with a space and sometimes not e.g. `variables/ columns` and `intercept/constant`, I propose you remove the additional spaces when used.
