### 2021-06-06 #3238 @Kevin-Mattheus-Moerman
Congratulations @tobiasschoch on getting this work published in JOSS!

Thanks @fboehm for editing this work!

Thanks also to @msalibian, @aalfons for the great job reviewing this work!
